获取scene相关信息的接口

+++ backend/springbootquickstart/src/main/java/com/easydeploy/springbootquickstart/model/ModelConfig.java
@@ -28,7 +28,7 @@ public class ModelConfig {
     private int batchSize;
     private double momentumValue;
     private double weightDecay;
-    private int sceneId;
+    private Long sceneId;
 
     @Enumerated(EnumType.STRING)
     private TrainingStatus status;
